refactor(profile): extract year to date arrow icon
