+++ projects/client/src/lib/components/icons/YearToDateArrow.svelte
